+++ maplibre-gl-js/src/geo/projection/covering_tiles.ts
@@ -288,6 +288,7 @@ export function coveringTiles(transform: IReadonlyTransform, options: CoveringTi
                 distanceToCenter3d,
                 transform.fov);
         }
+        thisTileDesiredZ = detailsProvider.getTileZoom?.(tileID, thisTileDesiredZ) ?? thisTileDesiredZ;
         thisTileDesiredZ = (options.roundZoom ? Math.round : Math.floor)(thisTileDesiredZ);
         thisTileDesiredZ = Math.max(0, thisTileDesiredZ);
         const z = Math.min(thisTileDesiredZ, maxZoom);