+++ maplibre-gl-js/src/geo/projection/projection_factory.ts
@@ -17,7 +17,7 @@ import type {Projection} from './projection.ts';
 import type {ITransform, TransformConstrainFunction} from '../transform_interface.ts';
 import type {ICameraHelper} from './camera_helper.ts';
 
-export function createProjectionFromName(name: ProjectionSpecification['type'], transformConstrain: TransformConstrainFunction | undefined, globalState: Record<string, any>): {
+export function createProjectionFromName(name: ProjectionSpecification['type'], transformConstrain: TransformConstrainFunction | undefined, globalState: Record<string, any>, specification?: ProjectionSpecification): {
     projection: Projection;
     transform: ITransform;
     cameraHelper: ICameraHelper;
@@ -34,10 +34,11 @@ export function createProjectionFromName(name: ProjectionSpecification['type'], 
     switch (name) {
         case 'equal-earth':
         {
+            const projection = new EqualEarthProjection(specification);
             return {
-                projection: new EqualEarthProjection(),
-                transform: new EqualEarthTransform(transformOptions),
-                cameraHelper: new EqualEarthCameraHelper(),
+                projection,
+                transform: new EqualEarthTransform(transformOptions, projection.parameters),
+                cameraHelper: new EqualEarthCameraHelper(projection.parameters),
             };
         }
         case 'mercator':