+++ maplibre-gl-js/src/shaders/shaders.ts
@@ -68,6 +68,7 @@ import projectionMercatorFrag from './glsl/_projection_mercator.fragment.glsl.g.
 import projectionGlobeVert from './glsl/_projection_globe.vertex.glsl.g.ts';
 import projectionGlobeFrag from './glsl/_projection_globe.fragment.glsl.g.ts';
 import projectionEqualEarthVert from './glsl/_projection_equal_earth.vertex.glsl.g.ts';
+import projectionEqualEarthFrag from './glsl/_projection_equal_earth.fragment.glsl.g.ts';
 import atmosphereFrag from './glsl/atmosphere.fragment.glsl.g.ts';
 import atmosphereVert from './glsl/atmosphere.vertex.glsl.g.ts';
 import skyFrag from './glsl/sky.fragment.glsl.g.ts';
@@ -122,7 +123,7 @@ export const shaders: {
     prelude: prepare(preludeFrag, preludeVert),
     projectionMercator: prepare(projectionMercatorFrag, projectionMercatorVert),
     projectionGlobe: prepare(projectionGlobeFrag, projectionGlobeVert),
-    projectionEqualEarth: prepare(projectionGlobeFrag, projectionEqualEarthVert),
+    projectionEqualEarth: prepare(projectionEqualEarthFrag, projectionEqualEarthVert),
     background: prepare(backgroundFrag, backgroundVert),
     backgroundPattern: prepare(backgroundPatternFrag, backgroundPatternVert),
     circle: prepare(circleFrag, circleVert),