+++ maplibre-style-spec/src/index.test.ts
@@ -84,6 +84,7 @@ function validSchema(k, v, obj, ref, version, kind) {
         'variableAnchorOffsetCollection',
         'sprite',
         'projectionDefinition',
+        'projectionTransition',
         'state',
         'fontFaces'
     ]);